Fix pad_sequence mixed dtype padding decomp (#184173)
Match the pad_sequence decomposition to native behavior by padding rows in the output dtype before scattering them into the output buffer.

Fixes #144080


Pull Request resolved: https://github.com/pytorch/pytorch/pull/184173
Approved by: https://github.com/oulgen

+++ torch/_decomp/decompositions.py
@@ -3168,7 +3168,7 @@ def pad_sequence(sequences, batch_first=False, padding_value=0.0, padding_side="
     out = sequences[0].new_full(out_dims, padding_value)
     dim_paddings = (0, 0) * len(trailing_dims)
     for i in range(sequences_size):
-        currseq = sequences[i]
+        currseq = sequences[i].to(dtype=out.dtype)
         pad_amount = max_len - currseq.size(0)
         if padding_side == "right":
             row = aten.constant_pad_nd(
